requirements.txt
streamlit>=1.25.0
pdfplumber>=0.7.5
docx2txt>=0.8
keybert>=0.7.0
sentence-transformers>=2.2.2
torch>=2.3.0
transformers>=4.30.0
language-tool-python>=2.8.0
numpy>=1.23.0
pandas>=1.5.0
scikit-learn
plotly>=5.15.0
fpdf>=1.7.2

